now we can handle device lost errors + fixed 32-bit build
`vl_graphics_end` now may return VL_GRAPHICS_SHOULD_TERMINATE which means the graphics context has been lost and you should reinitialize it to continue rendering

+++ Velours/ht.c
@@ -11,7 +11,7 @@ VL_HT_HASH(char) {
 }
 
 VL_HT_HASH(VlPtr) {
-	return (uint64_t) p;
+	return (uint64_t) (size_t) p;
 }
 
 VL_API char vl_ht_iterate(void *table, void **pos, VlHTEntry *entry) {
@@ -20,14 +20,14 @@ VL_API char vl_ht_iterate(void *table, void **pos, VlHTEntry *entry) {
 	VlHTHeader* header = VL_HT_HEADER(table);
 	size_t entry_size = (sizeof(uint64_t) + sizeof(char) + header->key_size + header->value_size);
 	size_t size = header->cap * entry_size;
-	size_t offset = (uint64_t) *pos - (uint64_t) table;
+	size_t offset = (size_t) *pos - (size_t) table;
 	while (offset < size) {
 		uint64_t *hash = (uint64_t*) ((char*) table + offset);
 		char *is_occupied = (char*)((char*)hash + sizeof(uint64_t));
 		void *key = (char*) is_occupied + sizeof(char);
 		void *value = (char*) key + header->key_size;
 		*pos = (char*) *pos + entry_size;
-		offset = (uint64_t) *pos - (uint64_t) table;
+		offset = (size_t) *pos - (size_t) table;
 		if (*is_occupied) {
 			entry->key = key;
 			entry->value = value;
